backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
piper-tts>=1.2.0
python-multipart>=0.0.6

# Authentication
sqlmodel>=0.0.14
aiosqlite>=0.19.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
pydantic-settings>=2.0.0
greenlet>=3.0.0
